Replace original image logo as svg
Replace original image logo as svg. Resize the logo svg to larger size
such that it can be better seen. Due to resizing, also shift the style
for the padding one level downward(from the container level to the
children level) as padding for the larger logo svg is smaller than that
of others.